Fix the case of wrong calculation of steps by modifying the jquery selector
The jquery selector used for getting the steps in the wizard
was changed as the older one (ie. "ul > li > a") matched
anchor tags anywhere inside the wizard container resulting in
wrong calculation of no. of steps. The changed selector filters
only those anchor tags that have the href attr starting with
'#step-'